refactor(graph_view): typed BranchChip + chips module (#77)
Move branch-chip construction out of graph_view/mod.rs into a new
chips.rs: optimize_branch_display, strip_remote, strip_prefix_in_labels,
chip_display_name, abbreviate_branch_label, and the REMOTE_ONLY_ICON /
SYNCED_ICON glyph consts.

optimize_branch_display now returns Vec<BranchChip> (label + style +
resolved branch name) instead of Vec<(String, Style)>. The former
render-time resolve_chip_branch call is folded into chip construction, so
each chip carries the real branch it resolves to; the render tail and
mouse hit-testing read chip.branch instead of re-deriving it. Resolution
uses the same inputs as before, so ChipHit targets are unchanged.

Tests moved to chips.rs and converted to assert on BranchChip fields
(label/style/branch) instead of tuple positions; one test added for the
folded branch resolution.
