+++ src/ui/graph_view/mod.rs
@@ -19,9 +19,13 @@ use crate::{
     pr::{CiStatus, PrContext, PrInfo, ReviewState},
 };
 
+mod chips;
 mod geometry;
 mod metrics;
 mod rows;
+
+pub use chips::REMOTE_ONLY_ICON;
+use chips::{optimize_branch_display, strip_remote};
 
 pub use geometry::{
     avatar_overlay_x, avatars_active, effective_graph_width, next_graph_cap, AVATAR_GAP_CELLS,
@@ -610,34 +614,6 @@ impl<'a> GraphViewWidget<'a> {
     }
 }
 
-/// Strip a real remote prefix (e.g. "origin/", "upstream/") from a ref,
-/// returning the bare branch name. Local branches that merely contain a slash
-/// ("feature/foo") are not remote refs.
-fn strip_remote<'n>(name: &'n str, remotes: &[String]) -> Option<&'n str> {
-    remotes.iter().find_map(|r| {
-        name.strip_prefix(r.as_str())
-            .and_then(|rest| rest.strip_prefix('/'))
-    })
-}
-
-/// Single source of truth for whether graph label chips drop the remote prefix:
-/// only when the repo has exactly one remote (the cloud icon then conveys
-/// remoteness, so `<remote>/` is redundant). Multi-remote repos keep prefixes to
-/// disambiguate which remote a ref belongs to.
-fn strip_prefix_in_labels(remotes: &[String]) -> bool {
-    remotes.len() == 1
-}
-
-/// The name shown on a branch chip: for a remote ref in a single-remote repo,
-/// the remote prefix is dropped; otherwise the full ref name is used.
-fn chip_display_name<'n>(name: &'n str, remotes: &[String]) -> &'n str {
-    if strip_prefix_in_labels(remotes) {
-        strip_remote(name, remotes).unwrap_or(name)
-    } else {
-        name
-    }
-}
-
 /// Nerd Font octicons for the open-PR badge and its actioned markers.
 const PR_BADGE_ICON: char = '\u{f407}'; // nf-oct-git_pull_request
 const PR_APPROVED_ICON: char = '\u{f42e}'; // nf-oct-check
@@ -759,229 +735,6 @@ fn pr_badge_color(pr: &PrInfo, theme: &Theme) -> Color {
     }
 }
 
-/// Nerd Font cloud glyph marking a branch that only exists on a remote.
-pub const REMOTE_ONLY_ICON: &str = "\u{f0c2}"; //
-/// Marks a local branch whose remote counterpart points at the same commit.
-const SYNCED_ICON: &str = "↔";
-
-/// Optimize branch name display
-/// - A lone remote ref gets a cloud icon; a local+remote pair collapses to one ↔ label
-/// - If a local branch matches its origin/xxx among other branches, show "xxx <-> origin"
-/// - Otherwise, show each name separately
-/// - Render in bold with the graph color, wrapped in brackets
-/// - Selected branch is shown with inverted colors
-fn optimize_branch_display(
-    branch_names: &[String],
-    is_head: bool,
-    color_index: usize,
-    selected_branch_name: Option<&str>,
-    theme: &Theme,
-    remotes: &[String],
-) -> Vec<(String, Style)> {
-    use std::collections::HashSet;
-
-    if branch_names.is_empty() {
-        return Vec::new();
-    }
-
-    // Max width for a single branch label (e.g., "[fix/feature-name]")
-    const MAX_LABEL_WIDTH: usize = 40;
-
-    // Split local and remote branches by real remote prefix (HashSet for O(1)
-    // lookup). A name is remote only when it starts with a configured remote
-    // (e.g. "origin/", "upstream/"), never merely for containing a slash.
-    let local_branches: HashSet<&str> = branch_names
-        .iter()
-        .filter(|n| strip_remote(n, remotes).is_none())
-        .map(|s| s.as_str())
-        .collect();
-
-    // Badge color always resolves through the same palette lookup as the
-    // graph line/node for this commit's lane (`theme.lane_color`), so a
-    // branch's badge and its line never diverge. HEAD is distinguished by
-    // weight (bold, below), not by a separate color.
-    let base_color = theme.lane_color(color_index);
-
-    // Helper to create style based on selection state. Restraint: color is the
-    // single emphasis device for ordinary chips; only the checked-out HEAD's
-    // chips also carry bold, reserving the stronger accent for the one ref that
-    // matters most. Selection adds REVERSED on top (orthogonal affordance).
-    let make_style = |branch_name: &str| -> Style {
-        let mut style = Style::default().fg(base_color);
-        if is_head {
-            style = style.add_modifier(Modifier::BOLD);
-        }
-        if selected_branch_name == Some(branch_name) {
-            // Reverse video rather than an explicit bg: when this branch's row
-            // is also the highlighted row, the list's highlight_style patches a
-            // bg over the whole line, which would clobber an explicit bg and
-            // leave the label invisible. REVERSED is resolved after that patch,
-            // so the selected branch stays legible on any terminal theme.
-            style.add_modifier(Modifier::REVERSED)
-        } else {
-            style
-        }
-    };
-
-    // Helper to create label with optional icon prefix and abbreviation
-    let make_label = |prefix: &str, name: &str, suffix: Option<&str>| -> String {
-        let prefix_width = display_width(prefix);
-        let (label, abbrev_width) = if let Some(s) = suffix {
-            (
-                format!("[{}{} {}]", prefix, name, s),
-                MAX_LABEL_WIDTH.saturating_sub(s.len() + 3 + prefix_width),
-            )
-        } else {
-            (
-                format!("[{}{}]", prefix, name),
-                MAX_LABEL_WIDTH.saturating_sub(prefix_width),
-            )
-        };
-
-        if display_width(&label) <= MAX_LABEL_WIDTH {
-            return label;
-        }
-
-        let abbrev = abbreviate_branch_label(name, abbrev_width, 0);
-        let abbrev = if prefix.is_empty() {
-            abbrev
-        } else {
-            abbrev.replacen('[', &format!("[{}", prefix), 1)
-        };
-        if let Some(s) = suffix {
-            abbrev.replace(']', &format!(" {}]", s))
-        } else {
-            abbrev
-        }
-    };
-
-    // Single-branch icon labels (VSCode Git Graph style): a lone remote ref
-    // gets a cloud icon; a local ref whose remote counterpart sits on the same
-    // commit collapses into one ↔-prefixed label. Commits carrying more than
-    // one distinct branch keep the standard rendering below.
-    if branch_names.len() == 1 {
-        let name = &branch_names[0];
-        if strip_remote(name, remotes).is_some() {
-            let prefix = format!("{} ", REMOTE_ONLY_ICON);
-            // Single-remote repos drop the "<remote>/" prefix (cloud conveys it).
-            let display = chip_display_name(name, remotes);
-            return vec![(make_label(&prefix, display, None), make_style(name))];
-        }
-    } else if branch_names.len() == 2 {
-        let synced_local = branch_names.iter().find(|name| {
-            strip_remote(name, remotes).is_none()
-                && branch_names
-                    .iter()
-                    .any(|other| strip_remote(other, remotes) == Some(name.as_str()))
-        });
-        if let Some(local) = synced_local {
-            let prefix = format!("{} ", SYNCED_ICON);
-            return vec![(make_label(&prefix, local, None), make_style(local))];
-        }
-    }
-
-    // Process branches in original order (matches tab order from filter_remote_duplicates)
-    let mut result: Vec<(String, Style)> = Vec::new();
-    for name in branch_names {
-        if let Some(bare) = strip_remote(name, remotes) {
-            // Remote branch: skip if matching local exists (dedup keeps a
-            // stripped remote-only chip from colliding with its local twin).
-            if local_branches.contains(bare) {
-                continue;
-            }
-            if strip_prefix_in_labels(remotes) {
-                // Single remote: drop the prefix but add the cloud icon so this
-                // remote-only chip still reads as remote in a multi-branch row.
-                let prefix = format!("{} ", REMOTE_ONLY_ICON);
-                result.push((make_label(&prefix, bare, None), make_style(name)));
-            } else {
-                result.push((make_label("", name, None), make_style(name)));
-            }
-        } else {
-            // Local branch: mark with the ↔ icon (same convention as the
-            // single-branch synced chip) when a remote ref points at the same
-            // bare name — dropping the redundant "↔ <remote>" text suffix.
-            let has_synced_remote = branch_names
-                .iter()
-                .any(|other| strip_remote(other, remotes) == Some(name.as_str()));
-            let prefix = if has_synced_remote {
-                format!("{} ", SYNCED_ICON)
-            } else {
-                String::new()
-            };
-            result.push((make_label(&prefix, name, None), make_style(name)));
-        }
-    }
-
-    // Collapse multiple branches: show up to two labels, then "+N" for the rest
-    if result.len() > 1 {
-        // Number of labels to display inline before collapsing the remainder
-        const SHOWN_LABELS: usize = 2;
-
-        let shown = SHOWN_LABELS.min(result.len());
-        let extra_count = result.len() - shown;
-
-        // Helper: split a formatted label into its leading icon prefix (cloud
-        // for a remote-only chip, ↔ for a synced local, or empty) and the bare
-        // branch name. The name is recovered for re-abbreviation; the icon is
-        // returned separately so the collapse below can re-attach it — a
-        // multi-branch row must keep its remote/synced markers, not drop them.
-        let split_label = |label: &str| -> (String, String) {
-            let s = label.trim_start_matches('[');
-            let (icon, rest) = if let Some(r) = s.strip_prefix(REMOTE_ONLY_ICON) {
-                (format!("{} ", REMOTE_ONLY_ICON), r.trim_start())
-            } else if let Some(r) = s.strip_prefix(SYNCED_ICON) {
-                (format!("{} ", SYNCED_ICON), r.trim_start())
-            } else {
-                (String::new(), s)
-            };
-            let bare = rest.split([']', ' ']).next().unwrap_or(label).to_string();
-            (icon, bare)
-        };
-
-        // Budget the available width across the shown labels
-        let per_label = MAX_LABEL_WIDTH / shown;
-
-        // Display order: always the stable badge order from `result` (which
-        // in turn follows `branch_names`, itself deterministically sorted at
-        // the source — see `BranchInfo::list_all`). Never reordered by which
-        // branch is selected/navigated to: that would make badge order flap
-        // as the cursor moves, which is the bug this fixes.
-        let mut combined = String::new();
-        for (pos, (label, _)) in result.iter().take(shown).enumerate() {
-            let (icon, clean_name) = split_label(label);
-            // Only the last shown label carries the "+N" suffix
-            let extra = if pos == shown - 1 { extra_count } else { 0 };
-            // Reserve budget for the icon so the abbreviated name plus its
-            // re-attached marker still fits the per-label allowance.
-            let budget = per_label.saturating_sub(display_width(&icon));
-            let abbrev = abbreviate_branch_label(&clean_name, budget, extra);
-            // Re-attach the icon marker inside the brackets (mirrors make_label).
-            let abbrev = if icon.is_empty() {
-                abbrev
-            } else {
-                abbrev.replacen('[', &format!("[{}", icon), 1)
-            };
-            combined.push_str(&abbrev);
-        }
-
-        // Style follows the selected branch when it's among these labels
-        // (highlight only — does not affect display order above).
-        let selected_idx = selected_branch_name
-            .and_then(|sel| {
-                branch_names
-                    .iter()
-                    .position(|n| n == sel || n.ends_with(&format!("/{}", sel)))
-            })
-            .unwrap_or(0)
-            .min(result.len().saturating_sub(1));
-        let style = result[selected_idx].1;
-        return vec![(combined, style)];
-    }
-
-    result
-}
-
 /// Determine which right-side elements (date, author, hash) to display based on available width.
 /// Returns (show_date, show_author, show_hash, total_right_width).
 ///
@@ -1033,61 +786,6 @@ fn compute_right_side_visibility(
 
     let right_width = width(show_date, show_author, show_hash);
     (show_date, show_author, show_hash, right_width)
-}
-
-/// Abbreviate branch name to max_width, showing "+N" if more branches exist
-/// Uses format: prefix/head...tail (preserving last 5 chars)
-fn abbreviate_branch_label(name: &str, max_width: usize, extra_count: usize) -> String {
-    const TAIL_LEN: usize = 5;
-    const ELLIPSIS: &str = "...";
-
-    let suffix = if extra_count > 0 {
-        format!(" +{}", extra_count)
-    } else {
-        String::new()
-    };
-
-    let suffix_len = display_width(&suffix);
-    let available = max_width.saturating_sub(suffix_len).saturating_sub(2); // -2 for brackets
-
-    // If name fits, return as-is
-    if display_width(name) <= available {
-        return format!("[{}]{}", name, suffix);
-    }
-
-    // Find "/" position to preserve prefix
-    let slash_pos = name.find('/');
-
-    // Split into prefix and rest
-    let (prefix, rest) = match slash_pos {
-        Some(pos) => (&name[..=pos], &name[pos + 1..]),
-        None => ("", name),
-    };
-
-    let prefix_width = display_width(prefix);
-    let ellipsis_width = display_width(ELLIPSIS);
-
-    // Get last TAIL_LEN characters from rest
-    let rest_chars: Vec<char> = rest.chars().collect();
-    let tail: String = if rest_chars.len() > TAIL_LEN {
-        rest_chars[rest_chars.len() - TAIL_LEN..].iter().collect()
-    } else {
-        rest.to_string()
-    };
-    let tail_width = display_width(&tail);
-
-    // Calculate available width for head portion
-    let head_available = available.saturating_sub(prefix_width + ellipsis_width + tail_width);
-
-    if head_available == 0 {
-        // Not enough space for head, just show truncated name
-        let truncated = truncate_to_width(name, available.saturating_sub(3));
-        return format!("[{}...]{}", truncated, suffix);
-    }
-
-    let head = truncate_to_width(rest, head_available);
-
-    format!("[{}{}{}{}]{}", prefix, head, ELLIPSIS, tail, suffix)
 }
 
 /// Format tag labels for a node. Tags render as `<name>` in the tag color,
@@ -1593,7 +1291,7 @@ fn render_graph_line_tail<'a>(
     let branch_width: usize = branch_display
         .iter()
         .enumerate()
-        .map(|(i, (label, _))| display_width(label) + if i > 0 { 1 } else { 0 })
+        .map(|(i, chip)| display_width(&chip.label) + if i > 0 { 1 } else { 0 })
         .sum::<usize>()
         + if !branch_display.is_empty() { 1 } else { 0 };
 
@@ -1628,33 +1326,35 @@ fn render_graph_line_tail<'a>(
     }
 
     // Render branch labels
-    for (i, (label, style)) in branch_display.iter().enumerate() {
+    for (i, chip) in branch_display.iter().enumerate() {
         if i > 0 {
             spans.push(Span::raw(" "));
             left_width += 1;
         }
         let chip_start = left_width;
-        left_width += display_width(label);
-        // Record a clickable region resolving to the underlying branch, so a
-        // click on the chip can check that branch out.
-        let branch = resolve_chip_branch(label, &node.branch_names, ctx.remotes);
-        let is_merged =
-            ctx.merged_dim && branch.as_deref().is_some_and(|n| ctx.merged_branches.contains(n));
-        if let Some(name) = branch {
+        left_width += display_width(&chip.label);
+        // Each chip already carries the branch a click on it resolves to (folded
+        // into chip construction, #77), so the render tail no longer re-derives it.
+        let is_merged = ctx.merged_dim
+            && chip
+                .branch
+                .as_deref()
+                .is_some_and(|n| ctx.merged_branches.contains(n));
+        if let Some(name) = &chip.branch {
             chips.push(ChipHit {
                 x_start: chip_start as u16,
                 x_end: left_width as u16,
-                target: ChipTarget::Branch(name),
+                target: ChipTarget::Branch(name.clone()),
             });
         }
         // A merged branch's chip keeps its lane hue, muted and dimmed (#90), so
         // it recedes while still reading as its own branch.
         let style = if is_merged {
-            merged_chip_style(*style, ctx.theme)
+            merged_chip_style(chip.style, ctx.theme)
         } else {
-            *style
+            chip.style
         };
-        spans.push(Span::styled(label.clone(), style));
+        spans.push(Span::styled(chip.label.clone(), style));
     }
     if !branch_display.is_empty() {
         spans.push(Span::raw(" "));
@@ -1743,37 +1443,6 @@ fn render_graph_line_tail<'a>(
     (Line::from(spans), chips)
 }
 
-/// Recover the branch name a rendered chip `label` refers to, matching it
-/// against the node's `branch_names`. Chip labels are decorated (`[name]`, an
-/// optional remote/synced icon prefix, a possible ` +N` overflow suffix), so we
-/// strip the decoration to a bare name and find the branch whose bare form
-/// matches (a local branch, or a remote ref bare-equal to it). Returns `None`
-/// when nothing matches (e.g. a non-branch decoration).
-fn resolve_chip_branch(label: &str, branch_names: &[String], remotes: &[String]) -> Option<String> {
-    // Strip the leading '[' and any icon prefix, then take up to the first
-    // delimiter (']', ' ', or the start of a "+N" overflow marker).
-    let s = label.trim_start_matches('[');
-    let s = s
-        .strip_prefix(REMOTE_ONLY_ICON)
-        .or_else(|| s.strip_prefix(SYNCED_ICON))
-        .map(str::trim_start)
-        .unwrap_or(s);
-    let bare = s.split([']', ' ']).next().unwrap_or(s);
-    if bare.is_empty() {
-        return None;
-    }
-    // Exact local match first, then a remote ref whose bare name matches.
-    branch_names
-        .iter()
-        .find(|n| n.as_str() == bare)
-        .or_else(|| {
-            branch_names
-                .iter()
-                .find(|n| strip_remote(n, remotes) == Some(bare))
-        })
-        .cloned()
-}
-
 impl<'a> StatefulWidget for GraphViewWidget<'a> {
     type State = ListState;
 
@@ -1812,87 +1481,6 @@ mod tests {
     use super::rows::fold_rows;
     use crate::pr::MergeState;
 
-    // ── optimize_branch_display icons ────────────────────────────────
-
-    fn labels(branch_names: &[&str], remotes: &[&str]) -> Vec<String> {
-        let names: Vec<String> = branch_names.iter().map(|s| s.to_string()).collect();
-        let remotes: Vec<String> = remotes.iter().map(|s| s.to_string()).collect();
-        let theme = Theme::dark();
-        optimize_branch_display(&names, false, 0, None, &theme, &remotes)
-            .into_iter()
-            .map(|(label, _)| label)
-            .collect()
-    }
-
-    #[test]
-    fn lone_remote_ref_gets_a_cloud_icon() {
-        // Single-remote repo: the "origin/" prefix is dropped (cloud conveys it).
-        let out = labels(&["origin/feature"], &["origin"]);
-        assert_eq!(out, vec![format!("[{} feature]", REMOTE_ONLY_ICON)]);
-    }
-
-    #[test]
-    fn lone_remote_ref_respects_non_origin_remotes() {
-        // Multi-remote repo: the prefix is kept to disambiguate the remote.
-        let out = labels(&["upstream/main"], &["origin", "upstream"]);
-        assert_eq!(out, vec![format!("[{} upstream/main]", REMOTE_ONLY_ICON)]);
-    }
-
-    #[test]
-    fn synced_local_and_remote_collapse_to_one_sync_label() {
-        let out = labels(&["main", "origin/main"], &["origin"]);
-        assert_eq!(out, vec![format!("[{} main]", SYNCED_ICON)]);
-    }
-
-    #[test]
-    fn synced_pair_order_does_not_matter() {
-        let out = labels(&["origin/main", "main"], &["origin"]);
-        assert_eq!(out, vec![format!("[{} main]", SYNCED_ICON)]);
-    }
-
-    #[test]
-    fn slashed_local_branch_is_not_mistaken_for_a_remote_ref() {
-        let out = labels(&["feature/foo"], &["origin"]);
-        assert_eq!(out, vec!["[feature/foo]".to_string()]);
-    }
-
-    #[test]
-    fn multiple_distinct_branches_get_no_icons() {
-        let out = labels(&["main", "dev"], &["origin"]);
-        for label in &out {
-            assert!(!label.contains(REMOTE_ONLY_ICON), "no cloud icon: {label}");
-            assert!(!label.contains(SYNCED_ICON), "no sync icon: {label}");
-        }
-    }
-
-    #[test]
-    fn two_unrelated_refs_do_not_collapse() {
-        // A local `main` and an unrelated remote `origin/dev` (no local twin):
-        // they must NOT merge into a single ↔ synced label, but the remote-only
-        // ref still carries its cloud marker (issue #74 — the cloud must not be
-        // dropped just because another chip shares the row).
-        let out = labels(&["main", "origin/dev"], &["origin"]);
-        assert!(
-            !out.iter().any(|l| l.contains(SYNCED_ICON)),
-            "unrelated local+remote must not be treated as synced: {out:?}"
-        );
-        let joined = out.join("");
-        assert!(
-            joined.contains(REMOTE_ONLY_ICON),
-            "remote-only ref keeps its cloud: {out:?}"
-        );
-        assert!(joined.contains("main") && joined.contains("dev"), "{out:?}");
-    }
-
-    #[test]
-    fn long_remote_ref_with_icon_stays_within_label_budget() {
-        let long = format!("origin/{}", "x".repeat(60));
-        let out = labels(&[&long], &["origin"]);
-        assert_eq!(out.len(), 1);
-        assert!(out[0].starts_with(&format!("[{} ", REMOTE_ONLY_ICON)));
-        assert!(display_width(&out[0]) <= 40, "label too wide: {}", out[0]);
-    }
-
     // ── build_tag_labels ─────────────────────────────────────────────
 
     #[test]
@@ -1929,95 +1517,6 @@ mod tests {
         assert!(display_width(label) <= 24, "label too wide: {label:?}");
         assert!(label.starts_with('<') && label.ends_with('>'));
         assert!(label.contains('…'), "expected an ellipsis: {label:?}");
-    }
-
-    // ── chip click resolution (resolve_chip_branch) ──────────────────────
-
-    #[test]
-    fn resolve_chip_branch_recovers_the_branch_name() {
-        let remotes = vec!["origin".to_string()];
-        let names = vec!["main".to_string(), "feature/x".to_string()];
-        // Plain local label.
-        assert_eq!(
-            resolve_chip_branch("[main]", &names, &remotes).as_deref(),
-            Some("main")
-        );
-        // Label with a slash in the name.
-        assert_eq!(
-            resolve_chip_branch("[feature/x]", &names, &remotes).as_deref(),
-            Some("feature/x")
-        );
-        // Synced-icon prefix is stripped before matching.
-        let synced = format!("[{} main]", SYNCED_ICON);
-        assert_eq!(
-            resolve_chip_branch(&synced, &names, &remotes).as_deref(),
-            Some("main")
-        );
-        // A cloud-icon remote-only chip (single-remote repo drops the prefix)
-        // resolves back to the full remote ref.
-        let remote_names = vec!["origin/dev".to_string()];
-        let cloud = format!("[{} dev]", REMOTE_ONLY_ICON);
-        assert_eq!(
-            resolve_chip_branch(&cloud, &remote_names, &remotes).as_deref(),
-            Some("origin/dev")
-        );
-        // No matching branch → None.
-        assert_eq!(resolve_chip_branch("[nope]", &names, &remotes), None);
-    }
-
-    // ── remote classification (strip_remote / optimize_branch_display) ───
-
-    #[test]
-    fn strip_remote_classifies_by_configured_remote_only() {
-        let remotes = vec!["origin".to_string(), "upstream".to_string()];
-        assert_eq!(strip_remote("upstream/main", &remotes), Some("main"));
-        assert_eq!(strip_remote("origin/feature/x", &remotes), Some("feature/x"));
-        // A slash alone does not make a branch remote.
-        assert_eq!(strip_remote("feature/x", &remotes), None);
-        assert_eq!(strip_remote("main", &remotes), None);
-        // Not a configured remote.
-        assert_eq!(strip_remote("fork/main", &remotes), None);
-    }
-
-    #[test]
-    fn multi_branch_dedupes_non_origin_remote_counterpart() {
-        let theme = Theme::dark();
-        // main + its upstream counterpart + an unrelated local branch. Three
-        // names skip the single/pair early-returns and hit the general loop.
-        let names = vec![
-            "main".to_string(),
-            "upstream/main".to_string(),
-            "dev".to_string(),
-        ];
-        let remotes = vec!["upstream".to_string()];
-        let out = optimize_branch_display(&names, false, 0, None, &theme, &remotes);
-        // Collapses to a single combined label.
-        assert_eq!(out.len(), 1);
-        let label = &out[0].0;
-        // upstream/main is recognized as main's remote counterpart and deduped,
-        // leaving two real branches — so no "+N" overflow marker appears.
-        assert!(
-            !label.contains('+'),
-            "non-origin remote counterpart should be deduped: {label:?}"
-        );
-        assert!(label.contains("main"), "expected main: {label:?}");
-        assert!(label.contains("dev"), "expected dev: {label:?}");
-    }
-
-    #[test]
-    fn slashed_local_branch_is_not_treated_as_remote() {
-        let theme = Theme::dark();
-        // "feature/x" contains a slash but "feature" is not a remote.
-        let names = vec!["feature/x".to_string()];
-        let remotes = vec!["origin".to_string()];
-        let out = optimize_branch_display(&names, false, 0, None, &theme, &remotes);
-        assert_eq!(out.len(), 1);
-        // No cloud icon (remote-only marker); rendered as a plain local label.
-        assert!(
-            !out[0].0.contains(REMOTE_ONLY_ICON),
-            "local branch must not get the remote icon: {:?}",
-            out[0].0
-        );
     }
 
     // ── open-PR badge matching ───────────────────────────────────────
@@ -2467,283 +1966,6 @@ mod tests {
             compute_right_side_visibility(60, cols(true, true, true)),
             (false, false, false, 0)
         );
-    }
-
-    // ── single-remote prefix stripping on labels ─────────────────────
-
-    #[test]
-    fn single_remote_strips_prefix_and_keeps_cloud_icon() {
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &["origin/feat".to_string()],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        let label = &out[0].0;
-        assert!(label.contains(REMOTE_ONLY_ICON), "cloud icon kept: {label:?}");
-        assert!(label.contains("feat"));
-        assert!(!label.contains("origin/"), "prefix stripped: {label:?}");
-    }
-
-    #[test]
-    fn upstream_only_remote_strips_prefix() {
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &["upstream/main".to_string()],
-            false,
-            0,
-            None,
-            &theme,
-            &["upstream".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        assert!(out[0].0.contains("main"));
-        assert!(!out[0].0.contains("upstream/"), "{:?}", out[0].0);
-    }
-
-    #[test]
-    fn multi_remote_keeps_prefix() {
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &["origin/feat".to_string()],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string(), "upstream".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        assert!(
-            out[0].0.contains("origin/feat"),
-            "multi-remote keeps prefix: {:?}",
-            out[0].0
-        );
-    }
-
-    #[test]
-    fn single_remote_synced_pair_still_collapses() {
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &["main".to_string(), "origin/main".to_string()],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string()],
-        );
-        // The local+remote pair collapses to one ↔ chip — no duplicate [main].
-        assert_eq!(out.len(), 1);
-        assert!(out[0].0.contains("main"));
-        assert!(out[0].0.contains(SYNCED_ICON), "synced icon: {:?}", out[0].0);
-        assert!(!out[0].0.contains("origin/"));
-    }
-
-    #[test]
-    fn single_remote_multi_branch_dedups_without_duplicate_or_prefix() {
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &[
-                "foo".to_string(),
-                "origin/foo".to_string(),
-                "bar".to_string(),
-            ],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        let label = &out[0].0;
-        // origin/foo is the remote twin of local foo → deduped, no duplicate and
-        // no leftover prefix; two real branches remain, so no "+N".
-        assert!(!label.contains("origin/"), "{label:?}");
-        assert!(!label.contains('+'), "no overflow marker: {label:?}");
-        assert!(label.contains("foo"));
-        assert!(label.contains("bar"));
-    }
-
-    #[test]
-    fn single_remote_remote_only_chip_in_multi_branch_resolves_to_name() {
-        let theme = Theme::dark();
-        // A remote-only ref alongside a local branch: the combined label must
-        // show the stripped name, never the bare cloud glyph.
-        let out = optimize_branch_display(
-            &["origin/lonely".to_string(), "bar".to_string()],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        let label = &out[0].0;
-        assert!(label.contains("lonely"), "stripped name present: {label:?}");
-        assert!(!label.contains("origin/"), "{label:?}");
-    }
-
-    // ── issue #74: multi-branch rows keep their cloud/synced markers ──
-    //
-    // Regression: the collapse path (result.len() > 1) rebuilt each chip from
-    // the cleaned bare name and dropped the icon prefix, so a row with two
-    // remote refs — or a synced pair alongside another branch — rendered as
-    // bare local-looking labels (`[mac][main]`). Assert the markers survive.
-
-    #[test]
-    fn two_remote_refs_on_one_commit_both_keep_cloud_icon() {
-        // The #74 repro: origin/mac + origin/main sit together on an older
-        // commit (no local ref there). Single remote → prefix dropped, but each
-        // chip must carry the cloud so it still reads as remote-only.
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &["origin/mac".to_string(), "origin/main".to_string()],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        let label = &out[0].0;
-        assert!(label.contains("mac") && label.contains("main"), "{label:?}");
-        assert!(!label.contains("origin/"), "prefix dropped: {label:?}");
-        // Two cloud glyphs — one per remote-only chip.
-        assert_eq!(
-            label.matches(REMOTE_ONLY_ICON).count(),
-            2,
-            "cloud on both remote chips: {label:?}"
-        );
-        assert!(!label.contains(SYNCED_ICON), "not synced: {label:?}");
-    }
-
-    #[test]
-    fn synced_pair_in_multi_branch_row_keeps_synced_icon() {
-        // A synced local+remote pair (main / origin/main) alongside an
-        // unrelated local branch: the pair collapses to one ↔ chip and the ↔
-        // marker must survive the multi-branch collapse.
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &[
-                "main".to_string(),
-                "origin/main".to_string(),
-                "dev".to_string(),
-            ],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        let label = &out[0].0;
-        assert!(label.contains(SYNCED_ICON), "synced marker kept: {label:?}");
-        assert!(label.contains("main") && label.contains("dev"), "{label:?}");
-        assert!(!label.contains("origin/"), "{label:?}");
-    }
-
-    #[test]
-    fn local_only_multi_branch_row_never_gets_a_cloud() {
-        // Two purely local branches with no remote counterparts: neither chip
-        // may acquire a cloud (or synced) marker.
-        let theme = Theme::dark();
-        let out = optimize_branch_display(
-            &["feature".to_string(), "hotfix".to_string()],
-            false,
-            0,
-            None,
-            &theme,
-            &["origin".to_string()],
-        );
-        assert_eq!(out.len(), 1);
-        let label = &out[0].0;
-        assert!(
-            !label.contains(REMOTE_ONLY_ICON),
-            "no cloud on local chips: {label:?}"
-        );
-        assert!(!label.contains(SYNCED_ICON), "no synced marker: {label:?}");
-        assert!(label.contains("feature") && label.contains("hotfix"), "{label:?}");
-    }
-
-    // ── badge order is stable regardless of selection (issue #50) ────
-
-    #[test]
-    fn badge_order_is_independent_of_which_branch_is_selected() {
-        // Three branches on one commit — more than SHOWN_LABELS (2), so this
-        // also exercises the collapse path. Regression: selecting each
-        // branch in turn (as branch-cycling navigation does) used to move
-        // that branch's chip to the front, flipping the visible order.
-        let theme = Theme::dark();
-        let names = [
-            "alpha".to_string(),
-            "beta".to_string(),
-            "origin/gamma".to_string(),
-        ];
-        let remotes = ["origin".to_string()];
-
-        let no_selection = optimize_branch_display(&names, false, 0, None, &theme, &remotes);
-        let selected_alpha =
-            optimize_branch_display(&names, false, 0, Some("alpha"), &theme, &remotes);
-        let selected_beta =
-            optimize_branch_display(&names, false, 0, Some("beta"), &theme, &remotes);
-        let selected_gamma =
-            optimize_branch_display(&names, false, 0, Some("origin/gamma"), &theme, &remotes);
-
-        // Only the label text (not the style/highlight) needs to stay fixed.
-        let text = |v: &[(String, Style)]| v.iter().map(|(s, _)| s.clone()).collect::<Vec<_>>();
-        assert_eq!(text(&no_selection), text(&selected_alpha));
-        assert_eq!(text(&no_selection), text(&selected_beta));
-        assert_eq!(text(&no_selection), text(&selected_gamma));
-    }
-
-    #[test]
-    fn badge_order_matches_source_order_two_labels() {
-        // Two non-synced branches render as one combined chip, e.g.
-        // "[mac][origin/mac]" (issue #50's exact example) — assert the
-        // bracket groups keep `branch_names`' order and never flip when a
-        // different branch becomes the "selected" one.
-        let theme = Theme::dark();
-        let names = ["mac".to_string(), "zzz-other".to_string()];
-        let out = optimize_branch_display(&names, false, 0, None, &theme, &[]);
-        assert_eq!(out.len(), 1);
-        assert_eq!(out[0].0, "[mac][zzz-other]");
-
-        // Selecting the second branch must not move it first.
-        let out_selected =
-            optimize_branch_display(&names, false, 0, Some("zzz-other"), &theme, &[]);
-        assert_eq!(out_selected[0].0, "[mac][zzz-other]");
-    }
-
-    // ── badge color matches lane color (issue #53) ────────────────────
-
-    #[test]
-    fn head_badge_color_matches_lane_color_not_a_fixed_head_color() {
-        // Regression: the checked-out HEAD branch's badge used to be forced
-        // to a fixed color (green) regardless of the commit's actual lane
-        // color, diverging from the graph line/node drawn in that lane.
-        let theme = Theme::dark();
-        for color_index in 0..theme.lane_colors.len() {
-            if color_index == crate::graph::colors::MAIN_BRANCH_COLOR {
-                continue; // main branch is blue either way — not the regression case
-            }
-            let out = optimize_branch_display(
-                &["feature".to_string()],
-                true, // is_head
-                color_index,
-                None,
-                &theme,
-                &[],
-            );
-            assert_eq!(out.len(), 1);
-            assert_eq!(
-                out[0].1.fg,
-                Some(theme.lane_color(color_index)),
-                "HEAD badge fg must equal the lane's own color at index {color_index}"
-            );
-        }
     }
 
     // ── unicode cell clamping ────────────────────────────────────────
